+++ server/src/entities/ConvToUser.ts
@@ -1,28 +1,40 @@
-import { Column, Entity, ManyToOne, PrimaryGeneratedColumn } from "typeorm";
+import { Field, ObjectType } from "type-graphql";
+import {
+  BaseEntity,
+  Column,
+  Entity,
+  ManyToOne,
+  PrimaryGeneratedColumn,
+} from "typeorm";
 import { Conversation } from "./Conversation";
 import { User } from "./User";
 
+@ObjectType()
 @Entity()
-export class ConvToUser {
+export class ConvToUser extends BaseEntity {
+  @Field()
   @PrimaryGeneratedColumn()
-  public convToUserId!: number;
+  convToUserId!: number;
 
-  @Column()
-  public conversationUuid!: string;
-
-  @Column()
-  public userEmail!: string;
-
+  @Field()
   @Column({ default: true })
-  public active!: boolean;
+  active!: boolean;
+
+  @Field()
+  @Column("uuid")
+  conversationUuid!: string;
+
+  @Field()
+  @Column()
+  userId!: number;
 
   @ManyToOne(() => Conversation, (conversation) => conversation.convToUsers, {
     cascade: false,
   })
-  public conversation!: Conversation;
+  conversation!: Conversation;
 
   @ManyToOne(() => User, (user) => user.convToUsers, {
     cascade: false,
   })
-  public user!: User;
+  user!: User;
 }